pyproject.toml
[tool.poetry]
name = "release-notes"
version = "0.1.0"
description = ""
authors = ["Константин Косолапов <knets@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
